+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 3.0.0
 
 * Offset next runtime with repair time taken - Issue #121
+* Add blocked status - Issue #284
 
 ## Version 2.0.6